// apps/api/src/modules/tenants/dto/update-tenant.dto.d.ts
declare class DireccionDto {
    departamento: string;
    municipio: string;
    complemento: string;
}
export declare class UpdateTenantDto {
    nombre?: string;
    nrc?: string;
    actividadEcon?: string;
    direccion?: DireccionDto;
    telefono?: string;
    correo?: string;
    nombreComercial?: string;
}
export {};
//# sourceMappingURL=update-tenant.dto.d.ts.map